Prettier creation of databases
Separated app.py to config.py and __init__.py
Added sql schema of database
Added better database creation method
