+++ .env.dev
@@ -1,4 +1,4 @@
-FLASK_APP=project/__init__.py
+FLASK_APP=server/__init__.py
 FLASK_ENV=development
 DATABASE_URL=postgresql://user:abcdef@postgres-db:5432/test
 SQL_HOST=postgres-db
