+++ Votum/app/app.py
@@ -2,25 +2,12 @@
 import os
 from flask import Flask, request
 from flask_sqlalchemy import SQLAlchemy
+from flask.cli import FlaskGroup
 from server.utils import send_file
+from server import app, db
 
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DATABASE_URL','postgresql://')
-db = SQLAlchemy(app)
-
-
-class User(db.Model):
-    __tablename__ = "users"
-
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(50), unique=True, nullable=False)
-    password = db.Column(db.String(50), nullable=False)
-    voted = db.Column(db.Boolean, nullable=False)
-
-    def __repr__(self):
-        return f'<User {self.username}>'
-
+cli = FlaskGroup(app)
+app.run(debug=True, host='0.0.0.0', port=5000)
 
 @app.route('/<path:path>')
 def static_file(path):
@@ -28,15 +15,10 @@ def static_file(path):
 
 @app.route('/')
 def root():
+    print("test")
     return send_file('index.html')
 
 
-@app.route("/db/create")
-def db_create():
-    db.drop_all()
-    db.create_all()
-    db.session.commit()
-    return 'Created the database'
 
 @app.post('/db/add')
 def db_add():
@@ -63,6 +45,4 @@ def login():
         }
     return { 'message': 'Login failed!' }
 
-
-if __name__ == '__main__':
-    app.run(debug=True, host='0.0.0.0', port=5000)
+
