+++ Votum/app/server/config.py
@@ -1,0 +1,9 @@
+import os
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+
+class Config(object):
+    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL", "sqlite://")
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
